renamed files for ease of use, and fixed middleware error handling
