+++ client/src/index.js
@@ -4,9 +4,9 @@ import './styles/index.css';
 import App from './components/App';
 import { Provider } from 'react-redux';
 import store from './store/index';
-import registerServiceWorker from './registerServiceWorker';
+// import registerServiceWorker from './registerServiceWorker';
 
 ReactDOM.render(<Provider store={store}>
                   <App />
                 </Provider>, document.getElementById('root'));
-registerServiceWorker();
+// registerServiceWorker();
